Move `BindingsBackedCodec.kt` to `:object-serialization`
With a potential impact to `gcc2speedscope` since generated type names are no
longer unpacked to avoid a dependency on `:base-services`.
