+++ platforms/core-configuration/configuration-cache/src/main/kotlin/org/gradle/configurationcache/serialization/codecs/JavaRecordCodec.kt
@@ -17,6 +17,9 @@
 package org.gradle.configurationcache.serialization.codecs
 
 import org.gradle.configurationcache.problems.PropertyKind
+import org.gradle.configurationcache.serialization.Decoding
+import org.gradle.configurationcache.serialization.Encoding
+import org.gradle.configurationcache.serialization.EncodingProducer
 import org.gradle.configurationcache.serialization.ReadContext
 import org.gradle.configurationcache.serialization.WriteContext
 import org.gradle.configurationcache.serialization.beans.readPropertyValue
